telegram-bot/requirements.txt
aiogram==3.15.0
qrcode==7.4.2
pillow==11.0.0
python-dotenv==1.0.1